Refactor: CSS - Split file
